+++ sdk/CHANGELOG.md
@@ -62,6 +62,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `client_without_tls` feature in favor of separate `client` and `tls` features;
 - `IncreaseNativeTokenSupplyOptions`;
 - `HARDENED` const;
+- `AliasIdDto` and `NftIdDto`;
 
 ## 0.3.0 - 2023-05-02